Move HeapSizeVerifier to guest staging
