+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/HeapSizeVerifier.java
@@ -22,16 +22,13 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.core.heap;
+package com.oracle.svm.guest.staging;
 
 import org.graalvm.word.UnsignedWord;
 import org.graalvm.word.impl.Word;
 
-import com.oracle.svm.guest.staging.GuestStagingDependencyBridge;
-import com.oracle.svm.guest.staging.SubstrateGCOptions;
+import com.oracle.svm.guest.staging.util.UserError;
 import com.oracle.svm.shared.util.SubstrateUtil;
-import com.oracle.svm.core.util.UserError;
-import com.oracle.svm.core.util.UserError.UserException;
 
 /**
  * Verifies that the heap size options are used consistently. Note that some checks seem redundant
@@ -88,11 +85,11 @@ public final class HeapSizeVerifier {
         }
     }
 
-    public static void verifyMinHeapSizeAgainstMaxAddressSpaceSize(UnsignedWord minHeapSize) throws UserException {
+    public static void verifyMinHeapSizeAgainstMaxAddressSpaceSize(UnsignedWord minHeapSize) {
         verifyAgainstMaxAddressSpaceSize(minHeapSize, MIN_HEAP_SIZE_NAME);
     }
 
-    public static void verifyMaxHeapSizeAgainstMaxAddressSpaceSize(UnsignedWord maxHeapSize) throws UserException {
+    public static void verifyMaxHeapSizeAgainstMaxAddressSpaceSize(UnsignedWord maxHeapSize) {
         verifyAgainstMaxAddressSpaceSize(maxHeapSize, MAX_HEAP_SIZE_NAME);
     }
 
@@ -120,7 +117,7 @@ public final class HeapSizeVerifier {
         }
     }
 
-    private static RuntimeException reportError(String message) throws UserException {
+    private static RuntimeException reportError(String message) {
         if (SubstrateUtil.HOSTED) {
             throw UserError.abort(message);
         }
